- Added better handling of potentially empty nutch directories

+++ portlet/catalog/catalog-portlet/src/main/java/org/nterlearning/crawl/nutch/CrawlTool.java
@@ -297,8 +297,10 @@ public class CrawlTool {
      */
     private void deleteURLTempFiles() {
         File urlFiles[] = new File(mURLDir.toString()).listFiles();
-        for (File urlFile : urlFiles) {
-            urlFile.delete();
+        if (urlFiles != null) {
+            for (File urlFile : urlFiles) {
+                urlFile.delete();
+            }
         }
     }
